requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-telegram-bot
apscheduler
pydantic-settings
openai